VIDSOL-682 refactor meeting room as reusable component - detekt fixes
