+++ app/src/main/java/com/vonage/android/screen/room/MeetingRoomScreenRoute.kt
@@ -35,6 +35,7 @@ import com.vonage.android.util.rememberNoiseSuppression
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.launch
 
+@Suppress("LongMethod")
 @Composable
 fun MeetingRoomScreenRoute(
     roomName: String,
