+++ vonage-meeting-room/src/main/java/com/vonage/android/meetingroom/MeetingRoomScreenTestTags.kt
@@ -1,0 +1,7 @@
+package com.vonage.android.meetingroom
+
+object MeetingRoomScreenTestTags {
+    const val MEETING_ROOM_TOP_BAR = "meeting_room_top_bar"
+    const val MEETING_ROOM_CONTENT = "meeting_room_content"
+    const val MEETING_ROOM_BOTTOM_BAR = "meeting_room_bottom_bar"
+}
